# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlobalChat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  clerkId: string;
  username: string;
  displayName: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  isBanned: boolean;
  /** @nullable */
  banReason?: string | null;
  createdAt: string;
}

export interface UserUpdate {
  username?: string;
  displayName?: string;
  avatarUrl?: string;
}

export interface UserList {
  users: User[];
  total: number;
}

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  memberCount: number;
  createdAt: string;
  isMember: boolean;
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  avatarUrl?: string;
}

export interface GroupUpdate {
  name?: string;
  description?: string;
  avatarUrl?: string;
}

export interface GroupList {
  groups: Group[];
}

export interface GroupMember {
  userId: string;
  groupId: number;
  user: User;
  joinedAt: string;
}

export interface GroupMemberList {
  members: GroupMember[];
  total: number;
}

export type MessageMessageType = typeof MessageMessageType[keyof typeof MessageMessageType];


export const MessageMessageType = {
  text: 'text',
  image: 'image',
  file: 'file',
  audio: 'audio',
  video: 'video',
} as const;

export interface Message {
  id: number;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  fileName?: string | null;
  messageType: MessageMessageType;
  senderId: string;
  sender: User;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  recipientId?: string | null;
  createdAt: string;
}

export type MessageInputMessageType = typeof MessageInputMessageType[keyof typeof MessageInputMessageType];


export const MessageInputMessageType = {
  text: 'text',
  image: 'image',
  file: 'file',
  audio: 'audio',
  video: 'video',
} as const;

export interface MessageInput {
  content?: string;
  fileUrl?: string;
  fileType?: string;
  fileName?: string;
  messageType: MessageInputMessageType;
}

export interface MessageList {
  messages: Message[];
  hasMore: boolean;
}

export interface Contact {
  id: number;
  userId: string;
  contactUserId: string;
  contact: User;
  createdAt: string;
}

export interface ContactInput {
  contactUserId: string;
}

export interface ContactList {
  contacts: Contact[];
}

export interface FileUploadRequest {
  fileName: string;
  fileType: string;
  fileSize?: number;
  /** Base64 encoded file content */
  fileData: string;
}

export interface UploadResult {
  url: string;
  fileType: string;
  fileName: string;
  fileSize: number;
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  bannedUsers: number;
  totalGroups: number;
  totalMessages: number;
  messagesToday: number;
  newUsersToday: number;
}

export interface AdminUserList {
  users: User[];
  total: number;
}

export interface BanInput {
  reason: string;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  user_joined: 'user_joined',
  user_banned: 'user_banned',
  user_unbanned: 'user_unbanned',
  group_created: 'group_created',
  message_deleted: 'message_deleted',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  actorName: string;
  /** @nullable */
  targetName?: string | null;
  createdAt: string;
}

export interface ActivityList {
  activities: ActivityItem[];
}

export type ListUsersParams = {
search?: string;
limit?: number;
offset?: number;
};

export type ListGroupMessagesParams = {
before?: number;
limit?: number;
};

export type ListDmMessagesParams = {
before?: number;
limit?: number;
};

export type ListAdminUsersParams = {
search?: string;
status?: ListAdminUsersStatus;
limit?: number;
offset?: number;
};

export type ListAdminUsersStatus = typeof ListAdminUsersStatus[keyof typeof ListAdminUsersStatus];


export const ListAdminUsersStatus = {
  all: 'all',
  active: 'active',
  banned: 'banned',
} as const;