# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlobalChat API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all registered users (for search / adding contacts)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups visible to the user
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupList"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a specific group with member count
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group (admin only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /groups/{groupId}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMember"
  /groups/{groupId}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Left group
  /groups/{groupId}/members:
    get:
      operationId: listGroupMembers
      tags: [groups]
      summary: List members of a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMemberList"

  # ─── Messages ─────────────────────────────────────────────────────────────────
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List saved contacts for private chat
      responses:
        "200":
          description: Contacts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
    post:
      operationId: addContact
      tags: [contacts]
      summary: Save a user as a contact to enable private chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{contactUserId}:
    delete:
      operationId: removeContact
      tags: [contacts]
      summary: Remove a saved contact
      parameters:
        - name: contactUserId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ─── Uploads ──────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users with ban status (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, banned]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Admin user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: Banned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{userId}/unban:
    post:
      operationId: unbanUser
      tags: [admin]
      summary: Unban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unbanned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        isBanned:
          type: boolean
        banReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, username, displayName, email, role, isBanned, createdAt]
...
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        isMember:
          type: boolean
      required: [id, name, memberCount, createdAt, isMember]
    GroupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        avatarUrl:
          type: string
      required: [name]
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        avatarUrl:
          type: string
    GroupList:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
      required: [groups]
    GroupMember:
      type: object
      properties:
        userId:
          type: string
        groupId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        joinedAt:
          type: string
          format: date-time
      required: [userId, groupId, user, joinedAt]
    GroupMemberList:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
        total:
          type: integer
      required: [members, total]
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        contactUserId:
          type: string
        contact:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, contactUserId, contact, createdAt]
    ContactInput:
      type: object
      properties:
        contactUserId:
          type: string
      required: [contactUserId]
    ContactList:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
      required: [contacts]
...
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
    BanInput:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
...